pyproject.toml
[project]
name = "ml-app"
version = "1.0.0"
description = "Sample ML pipeline app (Python 3.12 + uv + Docker + LightGBM + optuna + Ruff)"
requires-python = ">=3.12,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Bartosz Litwiniuk", email = "bartosz.litwiniuk@outlook.com"}
]
dependencies = [
    "pyyaml>=6.0",
    "rdata>=1.0.0",
    "requests>=2.31.0",
    "lightgbm>=3.3.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.2.2",
    "optuna>=0.10.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
    "integration: marks tests that require network (deselect with '-m \"not integration\"')",
]
addopts = [
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html"
]
log_cli = true
log_cli_level = "INFO"